v 1.0.4 完善了常用页面

+++ src/main/java/com/sagiri/blog/entity/Project.java
@@ -15,7 +15,7 @@ public class Project {
   /**
    * 项目的唯一标识符
    */
-  private long projectId;
+  private Integer projectId;
 
   /**
    * 项目的名称
